move get_subtrees to persisted_state_tree
